Backend/Services/requirements.txt
Flask
Flask-Cors
python-dotenv
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
pymysql
cryptography
requests
